Remove extra parens when no stacktrace available.
`weasel.repl/process-message` for `:eval-js` currently tries to *call* the string "No stacktrace available" when processing an error with no stacktrace. This results in a new error and the repl hanging. Instead it should just return the string as a value, which would match the behavior of `clojure.browser.repl/evaluate-javascript`.